(front) Header 수정 및 로그아웃 , 로그인 페이지 완료

+++ frontend/src/components/Layout/SidePanel.tsx
@@ -6,16 +6,18 @@ import { fetchAllFolder, findDefaultFolder } from "../../api/folder";
 import { useArchiveMutation } from "../../hooks/useArchiveMutation";
 import type { ArchiveResponse } from "../../types/archive";
 import FolderSelect from "../Folder/FolderSelect";
+import { useFolderStore } from "../../hooks/useFolderStore";
+import { useConfirmStore } from "../../hooks/useConfirmStore";
 
 interface SidePanelProps {
   isOpen: boolean;
   onClose: () => void;
   data: ArchiveResponse | null; // 수정 시 데이터
-  initialFolderId?: number | null; // ✅ 추가: 새로 만들 때의 초기 폴더 ID
 }
 
-const SidePanel = ({ isOpen, onClose, data, initialFolderId }: SidePanelProps) => {
+const SidePanel = ({ isOpen, onClose, data }: SidePanelProps) => {
     const {createArchive, updateArchive, deleteArchive, isSaving ,isDeleting } = useArchiveMutation(onClose);
+    const { activeFolder } = useFolderStore();
 
     // 1. 자동 입력을 위한 상태 관리
     const [url, setUrl] = useState('');
@@ -108,9 +110,9 @@ const SidePanel = ({ isOpen, onClose, data, initialFolderId }: SidePanelProps) =
         setKeywords([]);
         
         // ✅ 폴더 초기화 로직 보강
-        if (initialFolderId !== undefined && initialFolderId !== null && initialFolderId !== -1) {
+        if (activeFolder.id !== undefined && activeFolder.id !== null && activeFolder.id !== -1) {
           // 1순위: 현재 보고 있는 폴더가 있을 경우 그 폴더로 지정
-          setSelectedFolderId(initialFolderId);
+          setSelectedFolderId(activeFolder.id);
         } else if (folders) {
           // 2순위: 보고 있는 폴더가 없으면(전체보기 등) 기존 기본 폴더 로직 적용
           const defaultFolder = findDefaultFolder(folders);
@@ -119,7 +121,7 @@ const SidePanel = ({ isOpen, onClose, data, initialFolderId }: SidePanelProps) =
           }
         }
       }
-    }, [data, isOpen, folders ,initialFolderId]); // folders를 추가하여 데이터 로드 즉시 반영
+    }, [data, isOpen, folders ,activeFolder.id]); // folders를 추가하여 데이터 로드 즉시 반영
 
 
     // 2. AI 생성 버튼 핸들러
@@ -167,8 +169,18 @@ const SidePanel = ({ isOpen, onClose, data, initialFolderId }: SidePanelProps) =
 
     };
 
-    const handleDeleteInPanel = async (id:number) => {
-      await deleteArchive(id);
+    const confirm = useConfirmStore((state) => state.confirm);
+    const handleDeleteInPanel = async (e: React.MouseEvent ,id:number) => {
+      e.stopPropagation();
+      const isConfirmed = await confirm({
+        message: `해당 아카이브를 정말 삭제할까요?`,
+        confirmText: "삭제",
+        cancelText: "취소"
+      });
+
+      if (isConfirmed) {
+        deleteArchive(id);
+      } 
       onClose(); // 삭제 성공 시 패널 닫기 (커스텀 훅 내부 onSuccess에 넣거나 여기서 처리)
     };
 
@@ -315,7 +327,7 @@ const SidePanel = ({ isOpen, onClose, data, initialFolderId }: SidePanelProps) =
         </BtnSave>
         {data && (
         <BtnDelete 
-          onClick={()=> handleDeleteInPanel(data.id)} disabled={isDeleting}>
+          onClick={(e)=> handleDeleteInPanel(e,data.id)} disabled={isDeleting}>
                   🗑️
         </BtnDelete>
         )}
